test, refactor: 엔티티 수정으로 인한 테스트 코드 수정
