+++ src/test/java/com/manytooneserverjava/manitomember/ManitoMemberServiceTest.java
@@ -3,13 +3,13 @@ package com.manytooneserverjava.manitomember;
 import com.manytooneserverjava.manito.domain.Manito;
 import com.manytooneserverjava.manito.domain.ManitoRepository;
 import com.manytooneserverjava.manito.service.ManitoService;
-import com.manytooneserverjava.manito.web.ManitoCreateForm;
-import com.manytooneserverjava.manito.web.ManitoDto;
+import com.manytooneserverjava.manito.web.dto.ManitoCreateForm;
+import com.manytooneserverjava.manito.web.dto.ManitoDto;
 import com.manytooneserverjava.manitomember.domain.ManitoMember;
 import com.manytooneserverjava.manitomember.domain.ManitoMemberRepository;
 import com.manytooneserverjava.manitomember.service.ManitoMemberService;
-import com.manytooneserverjava.manitomember.web.GiftInfoForm;
-import com.manytooneserverjava.manitomember.web.ManitoMemberDto;
+import com.manytooneserverjava.manitomember.web.dto.GiftInfoForm;
+import com.manytooneserverjava.manitomember.web.dto.ManitoMemberDto;
 import com.manytooneserverjava.member.domain.Member;
 import com.manytooneserverjava.member.domain.MemberRepository;
 import jakarta.transaction.Transactional;
